Finalizando mock dos recursos

+++ codigo-fonte/scrum-back-end/src/main/java/br/ufes/dto/ItemBacklogProjetoDTO.java
@@ -20,7 +20,7 @@ public class ItemBacklogProjetoDTO {
 
 	private String descricao;
 
-	private Integer prioridade;
+	private Long prioridade;
 
 	private SituacaoItemProjetoEnum situacao;
